fix: stopRunning stops only the process if it is on foreground

+++ Kernel/lib.c
@@ -107,7 +107,7 @@ int regPrinting(void){
 void stopRunning(void) {
 	for(int i = 0; i < 20; i++){
 		TPCB * current = get_pcb_by_pid(i);
-		if(count_occurrences(current->semaphore->waiting_processes, 2) > 0){
+		if(current->scope == FOREGROUND && count_occurrences(current->semaphore->waiting_processes, 2) > 0){
 			kill_process(i);
 			return;
 		}
